Refactor player management and update ClassType enum
- Updated `ClassType` enum values and added new classes.
- Renamed `NewClient` packet to `CMsgNewClient` in `GamePackets.cs`.
- Modified `MapId` assignment in `MapObject.cs`.
- Removed default values for `Level`, `Experience`, and `Gold` in `PlayerEntityConfiguration.cs`.
- Changed mappings for `PointAllotData` and `MapData` from tables to views in `GameDbContext.cs`.
- Enhanced `PlayerMapper` to include additional properties for player creation.
- Refactored `SqlPlayerRepository` for clearer player creation and updating methods.
- Updated `MapData` property length and changed `Class` type in `PlayerData` to enum.
- Adjusted `PointAllotData` properties from `int` to `short`.
- Refactored `GameClient` to remove rate limiting and improve disconnect handling.
- Updated `PacketFactory` for accurate player property mapping.
- Refactored `GameServer` to utilize dedicated channels for clients.
- Improved encapsulation in `GameSystemsManager` with property usage.
- Enhanced player creation and updating logic in `PlayerManager`.
- Updated `app.manifest` for new application name and Windows compatibility.
- Added and updated migration files for database schema changes.
- Introduced a new `NewClient` handler for processing new client connections.

+++ MMORPGServer/Database/Models/PointAllotData.cs
@@ -14,16 +14,16 @@ namespace MMORPGServer.Database.Models
         public int Level { get; set; }
 
         [Column("force")]
-        public int Strength { get; set; }
+        public short Strength { get; set; }
 
         [Column("speed")]
-        public int Agility { get; set; }
+        public short Agility { get; set; }
 
         [Column("health")]
-        public int Vitality { get; set; }
+        public short Vitality { get; set; }
 
         [Column("soul")]
-        public int Spirit { get; set; }
+        public short Spirit { get; set; }
 
     }
 }
